Allow forcing tombstone dropping and targeting single tables in scrubs
The standalone scrubber tool got an option to force dropping tombstones when
performing an offline scrub.

The JMX scrub command got an option to pass in a list of SSTables to be
scrubbed, to allow targeted scrubs for tables that are known broken.

Support for targeted scrubs has also been added to the nodetool scrub command.